issue190: moved test out of driver package
